copy functionality across from the iframe bootstrapping code
